% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choixScores.R
\name{choixScores}
\alias{choixScores}
\title{choixScores}
\usage{
choixScores(
  factPattern,
  fctCorr = 1,
  meth = list("Thurstone", "Bartlett", "tenBerge"),
  N = 100
)
}
\arguments{
\item{factPattern}{todo}

\item{fctCorr}{todo}

\item{meth}{todo}

\item{N}{todo}
}
\value{
todo
}
\description{
choixScores
}
\author{
André Achim
}
