% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoeffMethode.R
\name{CoeffMethode}
\alias{CoeffMethode}
\title{CoeffMethode}
\usage{
CoeffMethode(factPattern, factCorr, nomMethode)
}
\arguments{
\item{factPattern}{todo}

\item{factCorr}{todo}

\item{nomMethode}{todo}
}
\value{
todo
}
\description{
CoeffMethode
}
\author{
André Achim
}
