% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribSources.R
\name{contribSources}
\alias{contribSources}
\title{contribSources}
\usage{
contribSources(factCoef, expanded, correlSources)
}
\arguments{
\item{factCoef}{(nv,nf) contient un jeu de coefficients factoriels}

\item{expanded}{(nv,nf+nv) donne la matrice d'amplitudes sur les facteurs (patrons) et les unicités}

\item{correlSources}{(nf+nv,nf+nv) est données par corrSources}
}
\value{
retourne une liste avec les coefficients factoriels tels que donnés en entrée, les mêmes ajustés pour donner une variance attendue des scores de 1.0 (facilite la comparaison entre méthodes) la corrélation attendue des scores et la contribution des nf+nv sources aux nf scores
}
\description{
contribSources
}
\author{
André Achim
}
