% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrSources.R
\name{corrSources}
\alias{corrSources}
\title{corrSources}
\usage{
corrSources(factPattern, factCorr = 1)
}
\arguments{
\item{factPattern}{todo}

\item{factCorr}{todo}
}
\value{
le résultat est (nf+nv, nf+nv).
}
\description{
corrélations entre toutes les sources, tant d'unicités que des facteurs.
}
\author{
André Achim
}
