% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_test.R
\name{factor_test}
\alias{factor_test}
\title{Title}
\usage{
factor_test(AS, AB)
}
\arguments{
\item{AS}{A SCFA object.}

\item{AB}{A pair c(i,j) to compare.}
}
\value{
Alist.
}
\description{
Title
}
\examples{
R <- Rnest::tabachnick_fidell2019
tf <- scfa(R, n = 175)
# Check for the pair 1 and 5
tf15 <- factor_test(AS = tf$AS, c(1,5))
}
