% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FP_methodes2.R
\name{FP_methodes2}
\alias{FP_methodes2}
\title{Todo}
\usage{
FP_methodes2(FS, N, ote, zero = NULL, nrep = 10)
}
\arguments{
\item{FS}{todo}

\item{N}{todo}

\item{ote}{todo}

\item{zero}{todo}

\item{nrep}{todo}
}
\value{
En sortie, on a les nombres de positifs (vrais ou faux). Les nombres de faux positifs sont donnée en négatif. Le première ligne imprimée est pour la corrélation partielle traditionnelle et la deuxième ligne pour partitionne_R.
}
\description{
Todo
}
\examples{
FS <- matrix(c(.7,.6,-.45,-.39,.5,.4,.3,0,0,0,0.6,.52,0,.3,.4,.5),ncol=2)
out <- FP_methodes2(FS, -500,c(1,2),c(2,3,6,9,13), nrep=1000)
out$sig


}
