% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\name{gen_data}
\alias{gen_data}
\title{Génération de jeux de données selon une structure factorielle.}
\usage{
gen_data(FS, N, R = NULL, with_var = FALSE, empirical = FALSE)
}
\arguments{
\item{FS}{Une structure factorielle.}

\item{N}{Nombre de sujets.}

\item{R}{une matrice de corrélation des facteurs.}

\item{with_var}{Si \code{with_var = TRUE}, la dernière colonne de F est la variance des unicités, autrement la variance d'unicité est 1 moins la somme des carrés sur les colonnes.}

\item{empirical}{logical. If \code{TRUE}, R specify the empirical not population covariance matrix.}
}
\value{
Une liste contenant un jeu de données (\code{dt}) et leur source \code{srce}.
}
\description{
Génération de jeux de données selon une structure factorielle.
}
