% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_SCA.R
\name{init_SCA}
\alias{init_SCA}
\title{Fonction pour initier SCA}
\usage{
init_SCA(R, N = NULL, seuils = c(0.001, 0.25))
}
\arguments{
\item{R}{une matrice de corrélation ou un jeu de données.}

\item{N}{si une matrice de corrélation est entrée, indiquer le nombre de sujets.}

\item{seuils}{Erreur de type I. Valeur défaut = .001 et .25.}
}
\value{
Retourne une liste avec plusieurs éléments de SCA.
}
\description{
Fonction pour initier SCA
}
\examples{
R <- Rnest::tabachnick_fidell2019
out <- init_SCA(R, N = 175)
}
