% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N1000_1.R
\docType{data}
\name{N1000_1}
\alias{N1000_1}
\title{An example for SignalCancellation}
\format{
A 1000 by 18 data frame.
}
\source{
\url{https://github.com/SignalCancellation}
}
\usage{
N1000_1
}
\description{
A data frame for SCRoF.
}
\keyword{datasets}
