% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N2000.R
\docType{data}
\name{N2000}
\alias{N2000}
\title{An example for SignalCancellation}
\format{
A 2000 by 18 data frame.
}
\source{
\url{https://github.com/SignalCancellation}
}
\usage{
N2000
}
\description{
A data frame for SCRoF.
}
\keyword{datasets}
