% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propScores.R
\name{propScores}
\alias{propScores}
\title{propScores}
\usage{
propScores(factCoef, factPattern, factCorr = 1, variances = 1)
}
\arguments{
\item{factCoef}{si donnée, est (nf,nf) (defaut I(nf,nf)).}

\item{factPattern}{(nv,nf) est la matrice de patrons factoriels.}

\item{factCorr}{si donnée, est (nf,nf) (defaut I(nf,nf)).}

\item{variances}{variances (optionnel) (nv) donne les variances des nv varables si autres que 1.}
}
\value{
TODO
}
\description{
TODO
}
\author{
André Achim
}
