% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_dim.R
\name{scan_dim}
\alias{scan_dim}
\alias{SCAN_DIM}
\alias{SCAN_Dim}
\title{Signal Cancellation Analysis of Number of Dimensions}
\usage{
scan_dim(
  .data,
  ...,
  n = NULL,
  alpha = 0.05,
  max.fact = NULL,
  missing = NULL,
  cluster = NULL,
  ordered = NULL,
  simulation = FALSE
)
}
\arguments{
\item{.data}{data frame, a numeric matrix, covariance matrix or correlation matrix from which to determine the number of factors.}

\item{...}{See details.}

\item{n}{the number of cases (subjects, participants, or units) if a covariance matrix is supplied in \code{.data}.}

\item{alpha}{a vector of type I error rates. Default is .05.}

\item{max.fact}{an optional maximum number of factor to extract. Default is \code{NULL}, so the maximum number possible.}

\item{missing}{todo}

\item{cluster}{todo}

\item{ordered}{todo}

\item{simulation}{FALSE}
}
\value{
A list.
}
\description{
\code{scan_dim} is used to identify the number of factors to retain in exploratory factor analysis.
}
\examples{
R <- Rnest::tabachnick_fidell2019
scan_dim(R, n = 175)
}
\author{
André Achim & P.-O. Caron
}
