% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreOrig.R
\name{scoreOrig}
\alias{scoreOrig}
\title{scoreOrig}
\usage{
scoreOrig(scoreProp, factCorr = 0)
}
\arguments{
\item{scoreProp}{todo}

\item{factCorr}{todo}
}
\value{
todo
}
\description{
scoreOrig
}
\author{
André Achim
}
