% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRoF.R
\name{SCRoF}
\alias{SCRoF}
\alias{SCROF}
\alias{scrof}
\title{SCRoF : Signal Cancellation Recovery of Factors}
\usage{
SCRoF(R, N = NULL, seuils = c(0.001, 0.25))
}
\arguments{
\item{R}{correlation, covariance matrix or a data set.}

\item{N}{sample size if correlation or covariance matrix is supplied.}

\item{seuils}{significance threshold.}
}
\value{
A list of class "SCROF"
}
\description{
SCRoF : Signal Cancellation Recovery of Factors
}
\examples{
res <- SCRoF(N1000_1)
}
\references{
Achim, A. (2024). Signal cancellation factor analysis. \doi{10.31234/osf.io/h7qwg}
Achim, A. (2025). Signal cancellation recovery of factors and meta-factors. \emph{Prooceedingd of the 89th Annual Meeting of the Psychometric Society}. Spring, Prague
}
\author{
Andr?? Achim (MATLAB)
Pier-Olivier Caron (R)
}
