% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionne_R.R
\name{partitionne_R}
\alias{partitionne_R}
\title{Partionne une matrice de corrélation en matrice de corrélation adimensionnelle}
\usage{
partitionne_R(R, N, ote)
}
\arguments{
\item{R}{A correlation matrix.}

\item{N}{Sample size.}

\item{ote}{vVcteur de identifiant des collonn, length(ote) doit être paire et >= 2.}
}
\value{
A matrice contenant les corrélations adimensionnelles dans le triangle inférieure et les valeurs p  dans le triangle supérieur.
}
\description{
Partionne une matrice de corrélation en matrice de corrélation adimensionnelle
}
\examples{
partionne_R(R = Rnest::tabachnick_fidell2019,
            N = 176,
            ote = c(2,3))
}
